requirements.txt
asgiref==3.4.1
Django==3.2.7
pip==21.2.4
pytz==2021.1
setuptools==58.0.4
sqlparse==0.4.2
djangorestframework==3.12.4
django-cors-headers==3.8.0
django-filter==21.1
mixer==7.1.2
drf-yasg==1.20.0
